Home view added. Snackbar added. Lazy loading in router modified. Searchbar extracted to single component. Minor adjustments
A Home view was created just for the sake of having a home instead of going directly to the Search view

A snackbar component was added to the app layout, together with methods for invoking it globally.

Lazy loading in the router is handled by a function for better readability.

The searchbar was extracted to a single component, which emits all events regarding the search itself.

Minor visual adjustments were made